fix(repositories): fix repositories relationship issues
Fix issues when trying to merge entities using JPA default save method

Issues: #1

+++ src/main/java/com/mcalzada/repository/CollaboratorRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CollaboratorRepository extends JpaRepository<Collaborator, Integer>
+public interface CollaboratorRepository extends JpaRepository<Collaborator, Long>
 {
 
     Optional<Collaborator> findFirstByName(String name);
